requirements.txt
discord.py>=2.0.0
yt-dlp>=2023.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
davey>=0.1.0